// syncro/apps/web/src/lib/api/generated/model/alertNotificationHistoryResponse.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Syncro API
 * Syncro backend REST API.
 * OpenAPI spec version: v1
 */
import type { NotificationJobView } from './notificationJobView';

export interface AlertNotificationHistoryResponse {
  items?: NotificationJobView[];
  total?: number;
}